Added more options and fixed printing of insert statements
Added option --verbose which saves as well as prints out data to the file
or console window. Added another option called --output-file which
specifies the name of the file to print the data. --assess-id specifies
the unique id for sql databases when just print is used statements.
Now just print also works with mongodb and prints out an array
of json formatted documents which can be imported into mongodb.